:art: refactor(PROBLEM2) : 유틸 클래스의 생성을 막기위해 비공개 생성자 추가 및 구조 개선
