+++ src/main/java/onboarding/problem2/domain/util/IndexCalculator.java
@@ -1,9 +1,14 @@
-package onboarding.problem2.domain;
+package onboarding.problem2.domain.util;
 
 import java.util.stream.IntStream;
 
 public class IndexCalculator {
+    private static final String UTILITY_CLASS_CREATE_EXCEPTION_MESSAGE = "생성할 수 없는 유틸리티 클래스입니다.";
     private static final int NOT_EXIST_OVERLAP_CHARACTER = -1;
+    
+    private IndexCalculator() {
+        throw new IllegalStateException(UTILITY_CLASS_CREATE_EXCEPTION_MESSAGE);
+    }
     
     public static int overlapStartIndex(final int overlapStartIndex, final String[] splitCryptogram) {
         return IntStream.range(overlapStartIndex, splitCryptogram.length - 1)
